{"schema":"decodex/commit/1","summary":"Split operator popover metrics by readout domain","authority":"manual"}
